Conflicts:
	src/main/java/com/kh/billida/member/model/repository/MemberRepository.java
	src/main/java/com/kh/billida/member/model/service/MemberService.java
	src/main/java/com/kh/billida/member/model/service/MemberServiceImpl.java

+++ src/main/java/com/kh/billida/rentLocker/model/dto/Locker.java
@@ -18,8 +18,8 @@ public class Locker {
 	private Date rentableDateEnd;
 	private String rentStatus;
 	private Long profit;
-	private Long latitude;
-	private Long longitude;
+	private float latitude;
+	private float longitude;
 	private Long rentedCnt;
 	
 }
